Backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.6.0
pandas>=2.1.0
python-multipart>=0.0.6
jinja2>=3.1.0
mangum>=0.17.0
python-dotenv>=1.0.0

